feat: correct sources.ts
